[ISSUE-4] Column count now uses python builtin method or input argument
Summary:
Previously, terminal size was computed using stty command. This did
not function cross-platform and was replaced with python's builtin
os.get_terminal_size() method. During this change, I also realized that
the testing environment was not agnostic to terminal size. To remedy this,
a new parameter for "number of columns in output" was added. The test's
golden results were created with column count of 80, so that value is not
explicitly passed in to the test commands that depend on it.

Tests:
All prior tests pass now on any terminal size.