[ISSUE-15] Allow gomp'ing against git tags
**Summary**
https://github.com/MarkForged/GOMP/issues/15
This commit adds the ability to use GOMP against git tags as well as branches. Tags may now be used in place of the left or right hand side.

**Test Plan**
Manually added tags to a branch in the GOMP repo, confirmed the updated script accepts tags as a parameter and will display the diff correctly.